#3 new navigation api implementation draft
