+++ widgets/src/androidMain/kotlin/dev/icerock/moko/widgets/screen/navigation/BottomNavigationScreen.kt
@@ -2,7 +2,7 @@
  * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.widgets.screen
+package dev.icerock.moko.widgets.screen.navigation
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -14,12 +14,17 @@ import android.widget.LinearLayout
 import androidx.core.view.ViewCompat
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import dev.icerock.moko.graphics.Color
+import dev.icerock.moko.widgets.screen.Args
+import dev.icerock.moko.widgets.screen.FragmentNavigation
+import dev.icerock.moko.widgets.screen.Screen
 import dev.icerock.moko.widgets.utils.ThemeAttrs
 import dev.icerock.moko.widgets.utils.dp
 
 actual abstract class BottomNavigationScreen actual constructor(
-    private val screenFactory: ScreenFactory
+    builder: BottomNavigationItem.Builder.() -> Unit
 ) : Screen<Args.Empty>() {
+    actual val items: List<BottomNavigationItem> = BottomNavigationItem.Builder().apply(builder).build()
+
     private val fragmentNavigation = FragmentNavigation(this)
     private var bottomNavigationView: BottomNavigationView? = null
 
@@ -51,7 +56,7 @@ actual abstract class BottomNavigationScreen actual constructor(
             item.icon?.also { menuItem.setIcon(it.drawableResId) }
 
             menuItemAction[menuItem] = {
-                val instance = screenFactory.instantiateScreen(item.screen)
+                val instance = item.screenDesc.instantiate()
                 fragmentNavigation.routeToScreen(instance)
             }
         }
@@ -95,7 +100,7 @@ actual abstract class BottomNavigationScreen actual constructor(
 
         if (savedInstanceState == null) {
             items.firstOrNull()?.let {
-                val instance = screenFactory.instantiateScreen(it.screen)
+                val instance = it.screenDesc.instantiate()
                 fragmentNavigation.setScreen(instance)
             }
         }
@@ -106,8 +111,6 @@ actual abstract class BottomNavigationScreen actual constructor(
 
         bottomNavigationView = null
     }
-
-    actual abstract val items: List<BottomNavigationItem>
 
     actual var selectedItemId: Int
         get() = bottomNavigationView?.selectedItemId ?: -1
