--- widgets/src/commonMain/kotlin/dev/icerock/moko/widgets/screen/DrawerNavigationScreen.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package dev.icerock.moko.widgets.screen
-
-import dev.icerock.moko.resources.ImageResource
-import dev.icerock.moko.resources.desc.StringDesc
-import kotlin.reflect.KClass
-
-expect abstract class DrawerNavigationScreen(
-    screenFactory: ScreenFactory
-) : Screen<Args.Empty> {
-    abstract val header: Any? // what info in header?
-    abstract val items: List<DrawerNavigationItem>
-
-    var selectedItemIndex: Int
-}
-
-data class DrawerNavigationItem(
-    val id: Int,
-    val title: StringDesc,
-    val icon: ImageResource? = null,
-    val screen: KClass<out Screen<Args.Empty>>
-)
